Refactor style/script functions into separate php partial
